requirements.txt
asttokens==2.4.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.5.0
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.25.0
jax==0.4.28
jax-cuda12-pjrt==0.4.28
jax-cuda12-plugin==0.4.28
jaxlib==0.4.28
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
ml-dtypes==0.4.0
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.5.40
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opt-einsum==3.3.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pgx==2.0.1
pgx-minatar==0.5.1
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
pyzmq==25.1.2
scipy==1.13.1
setuptools==78.1.1
six==1.16.0
stack-data==0.6.2
svgwrite==1.4.3
sympy==1.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5
traitlets==5.14.3
triton==3.3.1
typing_extensions==4.12.1
wcwidth==0.2.13
wheel==0.43.0
zipp==3.17.0
hydra-core
black
pytest
